requirements.txt
streamlit
nbformat
PyMuPDF
Pillow